+++ app/models/category.rb
@@ -3,7 +3,8 @@ class Category < ActiveRecord::Base
   extend FriendlyId
   friendly_id :title, use: :simple_i18n
   acts_as_list scope: :phase_id
-  include SocialDocument
+  include HasPopularity
+  include HasSubscriptors
 
   translates :title, :slug, :question, :body
   translation_required :title, :slug, :question