+++ app/models/concerns/has_popularity.rb
@@ -1,4 +1,15 @@
-module HasKudos
+# Measure model popularity
+# The model popularity has two types:
+# - Kudos - people saying I like it
+# - Page views - number of page views
+#
+module HasPopularity
+  extend ActiveSupport::Concern
+
+  included do
+    has_many :kudos, as: :document, dependent: :destroy
+  end
+
   def kudo(user)
     if user.class == User
       kudo = self.kudos.where(user_id: user.id).first
@@ -16,5 +27,10 @@ module HasKudos
         self.kudos.create!(ip: ip)
       end
     end
+  end
+
+  # This model has a view counter
+  def increment_view_counter
+    update_column(:view_count, self.view_count + 1)
   end
 end